refactor: move TryFrom<Value> for AstLiteral from ast to data
