+++ core/src/ast/mod.rs
@@ -6,7 +6,7 @@ mod function;
 mod operator;
 mod query;
 
-pub use ast_literal::{AstLiteral, AstLiteralError, DateTimeField, TrimWhereField};
+pub use ast_literal::{AstLiteral, DateTimeField, TrimWhereField};
 pub use data_type::DataType;
 pub use ddl::*;
 pub use expr::Expr;
